+++ tests/testthat/test-export-summs.R
@@ -287,6 +287,22 @@ test_that("plot.distributions works", {
                                   inner_ci_level = .9)))
 })
 
+if (requireNamespace("brms") & requireNamespace("broom.mixed")) {
+  bfit1 <- readRDS("brmfit.rds")
+  mvfit <- readRDS("mvfit.rds")
+  test_that("plot_coefs works with brms", {
+    expect_silent(print(plot_coefs(bfit1) + ggtitle("basic brms fit")))
+    expect_silent(print(plot_coefs(mvfit) + ggtitle("default mv brms fit")))
+    expect_silent(print(plot_coefs(mvfit, dpar = "sigma") +
+                          ggtitle("default dv, dpar sigma mv brms fit")))
+    expect_silent(print(plot_coefs(mvfit, resp = "wt", dpar = "sigma") +
+                          ggtitle("select wt dv, dpar sigma mv brms fit")))
+    expect_silent(print(plot_coefs(`MPG DV` = mvfit, `MPG Sigma` = mvfit, 
+                                   dpar = c(NA, "sigma"), resp = "mpg") +
+                          ggtitle("select mpg dv, dpar sigma separate models
+                                  mv brms fit")))
+  })
+}
 }
 
 options(device = device)